--- packages/gi-assets-basic/src/components/OperationPanel/index.md
@@ -1,3 +1,0 @@
-# 操作栏(OperationPanel)
-
-<code src="./doc/index.tsx"/>
