--- packages/gi-assets-basic/src/components/OperationPanel/index.tsx
@@ -1,10 +1,0 @@
-import Component from './Component';
-import info from './info';
-import registerMeta from './registerMeta';
-/**   index.md 中解析得到默认值，也可用户手动修改 */
-
-export default {
-  info,
-  component: Component,
-  registerMeta,
-};
