--- packages/gi-assets-basic/src/components/OperationPanel/info.ts
@@ -1,9 +1,0 @@
-const info = {
-  id: 'OperationPanel',
-  name: '操作项容器',
-  desc: '操作栏容器',
-  cover: 'http://xxxx.jpg',
-  category: 'container-components',
-  type: 'GICC',
-};
-export default info;
