+++ packages/gi-assets-basic/src/components/OperatorBar/Component.less
@@ -1,4 +1,4 @@
-.gi-operation-panel{
+.gi-operator-bar{
   position:absolute;
   display: flex;
   align-items: center;
@@ -6,11 +6,14 @@
   border: 1px solid #eee;
   box-sizing: border-box;
   overflow-y: hidden;
-  .gi-operation-panel-group{
+  box-shadow: var(--box-shadow-base);
+  .gi-operator-bar-group{
     position: relative;
     display: flex;
     align-items: center;
     justify-content: left;
+    background-color: var(--background-color);
+    height:100%;
     br{
       display: none;
       &+span{
@@ -18,7 +21,7 @@
         font-size: 12px !important;
       }
     }
-    .gi-operation-panel-group-driver{
+    .gi-operator-bar-group-driver{
       position: absolute;
       background-color: #ccc;
     }
