--- packages/gi-assets-basic/src/components/OperatorBar/index.less
@@ -1,9 +1,0 @@
-.gi-operator-bar {
-  position: absolute;
-  top: 0px;
-  right: 0px;
-  left: 0px;
-  width: 100%;
-  background-color: #ffffff;
-  box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.1);
-}
