+++ packages/gi-assets-basic/src/components/OperatorBar/index.md
