+++ packages/gi-assets-basic/src/components/OperatorBar/info.ts
@@ -1,0 +1,10 @@
+const info = {
+  id: 'OperatorBar',
+  name: '操作栏',
+  desc: '业务操作栏，可集成众多分析组件',
+  icon: 'icon-tabs',
+  cover: 'http://xxxx.jpg',
+  category: 'container-components',
+  type: 'GICC',
+};
+export default info;
