+++ packages/gi-assets-basic/src/components/index.tsx
@@ -15,7 +15,6 @@ import Loading from './Loading';
 import MiniMap from './MiniMap';
 import NeighborsQuery from './NeighborsQuery';
 import NodeLegend from './NodeLegend';
-import OperatorBar from './OperatorBar';
 import PathAnalysis from './PathAnalysis';
 import Placeholder from './Placeholder';
 import PropertiesPanel from './PropertiesPanel';
@@ -38,7 +37,7 @@ import ZoomIn from './ZoomIn';
 import ZoomOut from './ZoomOut';
 import ZoomStatus from './ZoomStatus';
 import RemoveNodeWithMenu from './RemoveNodeWithMenu';
-import OperationPanel from './OperationPanel';
+import OperatorBar from './OperatorBar';
 
 // 临时测试，后面要删掉，放到图谱业务资产库中
 // import FuseEditPanel from './FuseEditPanel';
@@ -88,5 +87,4 @@ export {
   ChartAnalysis,
   ZoomStatus,
   RemoveNodeWithMenu,
-  OperationPanel,
 };
