+++ packages/gi-site/src/pages/Analysis/MetaPanel/index.less
@@ -7,6 +7,11 @@
   .datav-gui .datav-gui-fill {
     border-radius: 2px !important;
   }
+
+  .ant-formily-array-collapse-item{
+    border:none;
+    background-color: var(--background-color);
+  }
 }
 
 // :root body{
